requirements.txt
flask 
flask-cors
pandas
numpy
bs4
newspaper3k
googletrans==3.1.0a0
datetime
gunicorn 
pymysql
sqlalchemy
googlesearch-python
spacy
dateparser
pyshorteners
python-decouple
secure-smtplib
pytest
requests_html
pytz
mypy
tabula-py
advertools
simpletransformers
click==8.0.4
pytest
luigi
pyglief
webdrive-manager